improved readability of nix_parser
