+++ nix_parser/src/grammar.rs
@@ -1,6 +1,4 @@
 use pyo3::prelude::*;
-use pyo3::types::PyAny;
-use pyo3::PyClass;
 
 use std::fmt;
